clinic/requirements.txt
django
djangorestframework
pillow
psycopg2==2.8.6
whitenoise==6.2.0
stripe
